requirements.txt
torch
torchvision
tifffile
numpy
scipy
matplotlib
Pillow
pyyaml
scikit-learn
imagecodecs
